editor/libeditor/tests/test_bug1130651.html

<!DOCTYPE html>
<title>Test for Bug 1130651</title>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1332876">Mozilla Bug 1332876</a>
<div contenteditable>a b</div>
<script>
var div = document.querySelector("div");
div.focus();
getSelection().collapse(div.firstChild, 2);
try {
  document.execCommand("inserttext", false, "\n");
  ok(true, "No exception thrown");
} catch (e) {
  ok(false, "Exception: " + e);
}
</script>
